solve the model names bug

+++ server_2/schemas/mutations/Account.js
@@ -3,11 +3,16 @@ const {
     GraphQLString, 
     GraphQLID 
 } = require("graphql");
-
+// import Account model
 const Account = require("../../database/models/Account");
+// import Account type
 const AccountType = require("../types/Account");
+// import Agent model
 const Agent = require("../../database/models/Agent");
+// import Client model
 const Client = require("../../database/models/Client");
+// import user authentication
+const authenticateUser = require("../auth");
 
 const AccountMutation = {
     // add account
@@ -19,7 +24,23 @@ const AccountMutation = {
             location: { type: new GraphQLNonNull(GraphQLString) },
             agent_id: { type: new GraphQLNonNull(GraphQLString) },
         },
-        async resolve(_parent, args) {
+        async resolve(_parent, args, context) {
+
+            // A list of models(users) that are allowed for this request
+            const allowedUsers = [ Admin,Agent ];
+
+            // authenticate the user
+            const authentiactedUser = await authenticateUser(allowedUsers, context);
+
+            // if user is authenticated
+            if (authentiactedUser === 1) {
+                throw new Error("User is not authentiacted.");
+            }
+
+            // if authenticated user is allowed for this request
+            if (authentiactedUser === 2 ) {
+                throw new Error("User is not authorized for this request.");
+            } 
 
             const account = new Account({
                 account_no: args.account_no,
@@ -46,7 +67,24 @@ const AccountMutation = {
         args: {
             id: { type: new GraphQLNonNull(GraphQLID) },            
         },
-        resolve(_parent, args) {
+        async resolve(_parent, args, context) {
+
+            // A list of models(users) that are allowed for this request
+            const allowedUsers = [ Admin ];
+
+            // authenticate the user
+            const authentiactedUser = await authenticateUser(allowedUsers, context);
+
+            // if user is authenticated
+            if (authentiactedUser === 1) {
+                throw new Error("User is not authentiacted.");
+            }
+
+            // if authenticated user is allowed for this request
+            if (authentiactedUser === 2 ) {
+                throw new Error("User is not authorized for this request.");
+            } 
+            
             return Account.findByIdAndDelete(args.id);
         }
     },
@@ -60,7 +98,25 @@ const AccountMutation = {
             location: { type: GraphQLString },
             agent_id: { type: GraphQLString },
         },
-        resolve(_parent, args) {
+        async resolve(_parent, args, context) {
+
+            // A list of models(users) that are allowed for this request
+            const allowedUsers = [ Admin, Agent, Client ];
+
+            // authenticate the user
+            const authentiactedUser = await authenticateUser(allowedUsers, context);
+
+            // if user is authenticated
+            if (authentiactedUser === 1) {
+                throw new Error("User is not authentiacted.");
+            }
+
+            // if authenticated user is allowed for this request
+            if (authentiactedUser === 2 ) {
+                throw new Error("User is not authorized for this request.");
+            } 
+
+            // more authorization checks here
             return Account.findByIdAndUpdate(
                 args.id,
                 {
