+++ server_2/schemas/mutations/Agent.js
@@ -3,9 +3,14 @@ const {
     GraphQLString, 
     GraphQLID 
 } = require("graphql");
-
+// import Agent
 const Agent = require("../../database/models/Agent");
-const { AgentType } = require("../types/Agent");
+// import Agent type and payload type
+const { AgentType, AgentAuthPayloadType } = require("../types/Agent");
+// import admin model
+const Admin = require("../../database/models/Admin");
+// import user authentication
+const authenticateUser = require("../auth");
 
 const AgentMutation = {
     // add agent
@@ -16,13 +21,58 @@ const AgentMutation = {
             email: { type: new GraphQLNonNull(GraphQLString) },
             phone: { type: new GraphQLNonNull(GraphQLString) },
         },
-        resolve(_parent, args) {
+        async resolve(_parent, args, context) {
+
+            // A list of models(users) that are allowed for this request
+            const allowedUsers = [ Admin ];
+
+            // authenticate the user
+            const authentiactedUser = await authenticateUser(allowedUsers, context);
+
+            // if user is authenticated
+            if (authentiactedUser === 1) {
+                throw new Error("User is not authentiacted.");
+            }
+
+            // if authenticated user is allowed for this request
+            if (authentiactedUser === 2 ) {
+                throw new Error("User is not authorized for this request.");
+            } 
+
             const agent = new Agent({
                 name: args.name,
                 email: args.email,
                 phone: args.phone,
             });
+            
             return agent.save();
+        }
+    },
+    // login agent
+    loginAgent: {
+        type: AgentAuthPayloadType,
+        args: { 
+            email: { type: new GraphQLNonNull(GraphQLString) },
+            password: { type: new GraphQLNonNull(GraphQLString) },
+        },
+        async resolve(_parent, args, _context) {
+
+            let agent = await Agent.exists({ email: args.email });
+            if (!agent) {
+                throw new Error("No such agent found");
+            }           
+
+            agent = await Agent.findById(agent._id);
+
+            const valid = await compare(args.password, agent.password);
+            if (!valid) {
+              throw new Error("Invalid cridentials");
+            }
+
+            const token = sign({agentId: agent.id}, process.env.APP_SECRET);
+            agent.token = token;
+
+            return agent;
         }
     },
     // delete agent
@@ -31,7 +81,24 @@ const AgentMutation = {
         args: {
             id: { type: new GraphQLNonNull(GraphQLID) },
         },
-        resolve(_parent, args) {
+        async resolve(_parent, args, context) {
+
+            // A list of models(users) that are allowed for this request
+            const allowedUsers = [ Admin ];
+
+            // authenticate the user
+            const authentiactedUser = await authenticateUser(allowedUsers, context);
+
+            // if user is authenticated
+            if (authentiactedUser === 1) {
+                throw new Error("User is not authentiacted.");
+            }
+
+            // if authenticated user is allowed for this request
+            if (authentiactedUser === 2 ) {
+                throw new Error("User is not authorized for this request.");
+            } 
+
             return Agent.findByIdAndRemove(args.id);
         },
     },
@@ -44,7 +111,24 @@ const AgentMutation = {
             email: { type: GraphQLString },
             phone: { type: GraphQLString },
         },
-        resolve(_parent, args) {
+        async resolve(_parent, args, context) {
+
+            // A list of models(users) that are allowed for this request
+            const allowedUsers = [ Admin ];
+
+            // authenticate the user
+            const authentiactedUser = await authenticateUser(allowedUsers, context);
+
+            // if user is authenticated
+            if (authentiactedUser === 1) {
+                throw new Error("User is not authentiacted.");
+            }
+
+            // if authenticated user is allowed for this request
+            if (authentiactedUser === 2 ) {
+                throw new Error("User is not authorized for this request.");
+            } 
+
             return Agent.findByIdAndUpdate(
                 args.id,
                 {
